Add cli script for local user creation
In order to make local dev environments work again, we need a way to
create a new user after the data reset.
Currently, the PlayerCreator assigns a generated password and sends a
registration-mail, unfortunately local email sending is a bit tricky -
so allow users to be created via cli and allow to define a password.